requirements.txt
fastapi[all]
SQLAlchemy
marshmallow-sqlalchemy
waitress
setuptools
requests
cachetools
mysqlclient
jsonify
PyJWT
